requirements.txt
numpy 
scikit-learn
pillow
flask
flask-cors
